refactor:import React생략, 리뷰 수정
