+++ babel.config.js
@@ -1,13 +1,9 @@
 module.exports = {
   presets: [
     [
-      '@babel/preset-env',
-      {
-        targets: {
-          node: 'current',
-        },
+      '@babel/preset-react', {
+        runtime: 'automatic',
       },
     ],
-    '@babel/preset-react',
   ],
 };
