+++ src/Todos.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import Todo from './Todo';
 
 function Todos({ todoItems, onClick }) {
